Migrate API routes from Pages Router to App Router
