--- pages/api/filings/enhanced-summary.ts
@@ -1,122 +1,0 @@
-/**
- * Enhanced SEC Filing Summary API
- * 
- * Provides SEC filing summaries with advanced features:
- * - Caching to prevent redundant API calls for the same filing
- * - Streaming support for faster partial results
- * - Enhanced chunking for large documents
- * - Improved error handling and recovery
- * 
- * @ts-ignore - Suppress TypeScript errors related to next-auth and getServerSession
- */
-
-import { NextApiRequest, NextApiResponse } from 'next';
-// Use type assertions to bypass import errors
-// @ts-ignore - Import errors will be fixed when next-auth is properly installed
-import { getServerSession } from 'next-auth/next';
-// @ts-ignore - Import errors will be fixed when auth options are properly defined
-import { authOptions } from '../auth/[...nextauth]';
-import { enhancedFilingService } from '../../../services/enhancedFilingService';
-import { logger } from '../../../lib/logging';
-import { FilingType } from '../../../lib/sec-edgar/types';
-import { getFormMetadata } from '../../../lib/sec-edgar/form-registry';
-import { prisma } from '../../../lib/db';
-
-// API route logger
-const apiLogger = logger.child('api-enhanced-summary');
-
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const startTime = Date.now();
-  
-  try {
-    // Validate request method
-    if (req.method !== 'GET') {
-      return res.status(405).json({ error: 'Method not allowed' });
-    }
-    
-    // Get user session
-    // @ts-ignore - Using three arguments version for compatibility with next-auth
-    const session = await getServerSession(req, res, authOptions);
-    const userId = session?.user?.id;
-    
-    // Extract query parameters
-    const { ticker, formType, useStreaming, useCache, processAllChunks } = req.query;
-    
-    // Validate required parameters
-    if (!ticker || typeof ticker !== 'string') {
-      return res.status(400).json({ error: 'Ticker is required' });
-    }
-    
-    if (!formType || typeof formType !== 'string') {
-      return res.status(400).json({ error: 'Form type is required' });
-    }
-    
-    // Validate form type
-    const formMetadata = getFormMetadata(formType as FilingType);
-    if (!formMetadata) {
-      return res.status(400).json({ error: `Invalid form type: ${formType}` });
-    }
-    
-    // Parse options
-    // @ts-ignore - Suppressing TypeScript error on line 63
-    const options = {
-      userId,
-      useStreaming: useStreaming === 'true',
-      useCache: useCache !== 'false', // Default to true
-      processAllChunks: processAllChunks === 'true'
-    };
-    
-    // Log request
-    apiLogger.info(`Enhanced summary request: ${ticker} ${formType}`, { options });
-    
-    // Record request in analytics
-    if (userId) {
-      // @ts-ignore - userActivity might not be in the Prisma schema types yet
-      await prisma.userActivity.create({
-        data: {
-          userId,
-          activityType: 'FILING_SUMMARY_REQUEST',
-          details: {
-            ticker,
-            formType,
-            options
-          }
-        }
-      });
-    }
-    
-    // Get filing summary
-    const result = await enhancedFilingService.getFilingSummary(
-      ticker.toUpperCase(),
-      formType as FilingType,
-      options
-    );
-    
-    // Handle error
-    if (!result.data) {
-      return res.status(404).json({ error: result.error || 'Filing summary not found' });
-    }
-    
-    // Return success response
-    // @ts-ignore - Suppressing TypeScript error on line 101
-    return res.status(200).json({
-      success: true,
-      data: result.data,
-      processingTimeMs: Date.now() - startTime
-    });
-    
-  } catch (error) {
-    // Log error
-    apiLogger.error('Error in enhanced summary API', { error });
-    
-    // Handle known API errors
-    if (error instanceof Error && 'statusCode' in error) {
-      const apiError = error as Error & { statusCode: number };
-      return res.status(apiError.statusCode).json({ error: apiError.message });
-    }
-    
-    // Handle unknown errors
-    const message = error instanceof Error ? error.message : 'Unknown error';
-    return res.status(500).json({ error: `Failed to get filing summary: ${message}` });
-  }
-}
